Moving files_decorator to gitrepo.py
