+++ datalad/support/annexrepo.py
@@ -18,7 +18,7 @@ import logging
 
 from ConfigParser import NoOptionError
 
-from gitrepo import GitRepo
+from gitrepo import GitRepo, files_decorator
 from datalad.cmd import Runner as Runner
 from exceptions import CommandNotAvailableError, CommandError, FileNotInAnnexError, FileInGitError
 
@@ -43,65 +43,6 @@ def _options_decorator(func):
         return func(self, *args, options=option_list)
     return newfunc
 
-def _normalize_path(base_dir, path):
-        """Helper to check paths passed to methods of this class.
-
-        Checks whether `path` is beneath `base_dir` and normalize it. Additionally paths are converted into
-        relative paths with respect to `base_dir`, considering os.getcwd() in case of relative paths.
-        This is intended to be used in repository classes, which means that `base_dir` usually will be
-        the repository's base directory.
-
-        Parameters:
-        -----------
-        path: str
-            path to be normalized
-        base_dir: str
-            directory to serve as base to normalized, relative paths
-
-        Returns:
-        --------
-        str:
-            normalized path, that is a relative path with respect to `base_dir`
-        """
-        path = normpath(path)
-        if isabs(path):
-            if commonprefix([path, base_dir]) != base_dir:
-                raise FileNotInAnnexError(msg="Path outside repository: %s" % path, filename=path)
-            else:
-                pass
-
-        elif commonprefix([getcwd(), base_dir]) == base_dir:
-            # If we are inside repository, rebuilt relative paths.
-            path = p_join(getcwd(), path)
-        else:
-            # We were called from outside the repo. Therefore relative paths
-            # are interpreted as being relative to self.path already.
-            return path
-
-        return relpath(path, start=base_dir)
-
-def _files_decorator(func):
-    """Decorator to provide unified path conversions.
-
-    Note: This is intended to be used within the repository classes and therefore returns
-    a class method!
-    The decorated function is expected to take a path or a list of paths at first positional
-    argument (after 'self'). Additionally the class `func` is a member of, is expected to have
-    an attribute 'path'.
-    """
-
-    def newfunc(self, files, *args, **kwargs):
-        if isinstance(files, basestring):
-            files_new = _normalize_path(self.path, files)
-        elif isinstance(files, list):
-            files_new = []
-            for path in files:
-                files_new.append(_normalize_path(self.path, path))
-        else:
-            raise ValueError("_files_decorator: Don't know how to handle instance of %s." %
-                             files.__class__.__name__)
-        return func(self, files_new, *args, **kwargs)
-    return newfunc
 
 
 class AnnexRepo(GitRepo):
@@ -228,7 +169,7 @@ class AnnexRepo(GitRepo):
             lgr.error('git annex init returned status %d.' % status)
 
     @_options_decorator
-    @_files_decorator
+    @files_decorator
     def annex_get(self, files, options=[]):
         """Get the actual content of files
 
@@ -258,7 +199,7 @@ class AnnexRepo(GitRepo):
             raise CommandError(cmd=' '.join(cmd_list))
 
     @_options_decorator
-    @_files_decorator
+    @files_decorator
     def annex_add(self, files, options=[]):
         """Add file(s) to the annex.
 
@@ -315,7 +256,7 @@ class AnnexRepo(GitRepo):
 
         return output
 
-    @_files_decorator
+    @files_decorator
     def get_file_key(self, path):
         """Get key of an annexed file
 
@@ -364,7 +305,7 @@ class AnnexRepo(GitRepo):
 
         return key
 
-    @_files_decorator
+    @files_decorator
     def file_has_content(self, path):
         """ Check whether the file `path` is present with its content.
